requirements.txt
streamlit
pandas
regex
pytesseract
Pillow
torch
transformers
tokenizers
sentencepiece
protobuf
requests
